pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Christoph Minixhofer", email = "christoph.minixhofer@gmail.com"},
]
dependencies = [
    "pytorch-lightning>=1.5.10",
    "torchaudio>=0.11.0",
    "wandb>=0.12.11",
    "pillow>=9.0.1",
    "matplotlib>=3.5.1",
    "scikit-learn>=1.0.2",
    "seaborn>=0.11.2",
    "phones>=0.0.2",
    "pandarallel>=1.5.7",
    "tgt>=1.4.4",
    "pyworld>=0.3.0",
    "llvmlite>=0.38.0",
    "librosa>=0.9.1",
    "click>=8.0.4",
    "diskcache>=5.4.0",
    "torchvision>=0.12.0",
    "plotly>=5.7.0",
    "kaleido==0.2.1",
    "snreval>=1.2",
    "audiomentations>=0.24.0",
    "black>=22.3.0",
    "crepe>=0.0.12",
    "textgrid>=1.5",
    "rich>=12.4.4",
    "alignments>=0.0.3",
    "g2p-en>=2.1.0"]
license = {text = "MIT"}
requires-python = ">=3.8,<3.10"
[project.urls]
Homepage = ""

[project.optional-dependencies]
dev = [
    "line-profiler>=3.5.1",
    "viztracer>=0.15.3",
]
[tool.pdm]
[tool.pdm.dev-dependencies]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
